treectrl update to 2.4.2+ for possible crash fix. The following sequence was reliably causing a crash. set o [wits::app::get_objects ::wits::app::user]
Start the user list view from the gui

Then in console
$o get_field

which should result in wrong # args message.
Instead it would crash or the error message would
be garbled.
Some investigation showed crash went away as the list view
update were commented out. So updated to current treectrl
from SF and that seemed to fix the crash. Still not
clear what changed in treectrl that fixed this or
 if the crash has simply moved elsewhere.